+++ src/repository/xtream_playlist_iterator.rs
@@ -40,7 +40,7 @@ impl XtreamPlaylistIterator {
                 .map_err(|err| info_err!(format!("Could not deserialize file {xtream_path:?} - {err}")))?;
 
             let options = XtreamMappingOptions::from_target_options(target, xtream_output, config);
-            let server_info = config.get_user_server_info(user).await;
+            let server_info = config.get_user_server_info(user);
 
             let filter = user_get_bouquet_filter(config, &user.username, category_id, TargetType::Xtream, cluster).await;